fix review directory structure
