+++ backend/review/views.py
@@ -13,22 +13,7 @@ from .models import Review
 from tensorflow.keras import models
 from konlpy.tag import Okt
 import nltk
-import numpy as np
-
-
-def tokenize(doc, okt):
-    return ['/'.join(t) for t in okt.pos(doc, norm=True, stem=True)]
-
-def term_frequency(doc, selected_words):
-    return [doc.count(word) for word in selected_words]
-
-def predict_score(model, okt, selected_words, review):
-    token = tokenize(review, okt)
-    tf = term_frequency(token, selected_words)
-    data = np.expand_dims(np.asarray(tf).astype('float32'), axis=0)
-    score = float(model.predict(data))
-
-    return round(score,2) * 10
+from . import sentiment as sentiment
 
 
 def user_review(request, reviewee_id):
@@ -71,7 +56,7 @@ def user_review(request, reviewee_id):
             text = nltk.Text(tokens, name='NMSC')
             okt = Okt()
             selected_words = [f[0] for f in text.vocab().most_common(10000)]
-            rating = predict_score(model, okt, selected_words, content)
+            rating = sentiment.predict_score(model, okt, selected_words, content)
 
             new_review = Review(reviewee=reviewee_user,
                                 rating=rating,
